Allow empty suites/functionalSuites command-line args
Also ensure that the state of `args.suites` is used when determining
whether or not to run ClientSuite in the runner, since that argument
is passed to the client and decides which unit test suites execute.

Fixes #268. Closes #269.

(cherry picked from commit be9753ceb2cc885420caca2bdfbabadb60f40ca6)

Conflicts:
	runner.js