fix duplicate suite reports with junit reporter
When the junit reporter is run with the runner, the reports are
duplicated. It appears that this is caused by the `/suite/end` topic
being published by both `Suite` and `createProxy` in
`publishInSequence`. This appears to only be an issue with the junit
reporter, as it appears to count on `/suite/end` to only be called once
per suite.