Improve client suite communication
Information no longer is transmitted directly through the reporter
manager for stuff coming from the client; this fixes several problems
with events from the client passing through to the runner that never
should pass through (like the suiteEnd event for the client-side
ClientSuite suite). It also fixes an infinite loop when the runner
is interrupted while ClientSuite is running; in this case, the
suiteEnd message from the ClientSuite was broadcast, picked up by
its own reporter that it added to the global reporter manager, causing
it to add itself to its own set of tests, resulting in infinite loop.