Replace half-baked unitTestWaitForRunner with runnerClientReporter
This allows the client-side reporter used by the test runner for
communication to be configured or replaced with a custom reporter.

This also changes the behaviour versus earlier versions of Intern 3
to avoid a performance regression. Users can now opt-in to unit
test waiting by setting `runnerClientReporter.waitForRunner` to
`true` or `'fail'`.