Pass a context-specific module loader when loading tunnels
The `dojo` package is mapped for `intern` only, so using the global
`require` in this context causes the loader to look in an incorrect
location for it. In other cases we still want to use the global
`require`, so passing a context-specific loader function to
lib/util.getModule is optional.

Fixes #460.

(cherry picked from commit 071cc6a7355e58923e3f36072eb5f52ce7db9712)