Proxy should not wait for test system when the client does not need it
Since the client does not care how the server responds in this
case, the server should simply respond immediately. If the server
does not respond immediately, and the browser sends more than
maxConnectionsPerServer messages out-of-order (which IE does),
all of the browser’s available connections may be exhausted and
the testing system will stall.

Refs #466.