Fall back to copying things to node_modules/ if symlinking fails.
Symlinking fails on a Windows machine when installing with npm 3.x where the intern install directory is on a Samba share (EUNKNOWN, the underlying win32 error is "NOT REPARSE POINT" from FileSystemControl). @csnover mentions in https://github.com/theintern/intern/issues/500#issuecomment-142804696 that falling back to copying is an acceptable option if symlinking fails.

Uninstalling with `npm uninstall` appears to delete the intern/node_modules directory and the intern/ directory, so there's no need for a separate uninstall script to delete it manually.