Handle globs in suite lists
Glob entries of the form 'package/dir/**/*' are now resolved to lists of
module IDs. In a Node.js context, Intern will resolve the package name
to its filesystem path and use the `glob` package to find all matching
modules. These modules will be converted back to module IDs and added to
the Intern config. Both `suites` and `functionalSuites` are handled this
way.

In the browser context, the `suites` list is now always sent back to a
special '__resolveSuites__' endpoint on Intern's proxy for resolution.
The list is resolved in the same manner as in the Node client, and the
resolved suites list will be returned to the browser and added to
Intern's config.

When `client.html` is used without Intern's proxy, or another server
providing data for the '__resolveSuites__' request, the request will
fail and a warning will be logged to the browser console stating that
resolution will not be performed. This will only affect the testing
process if glob expressions were actually used in any of the suite
names.

Resolution happens in PreExecutor just after loader swapping, so loader
maps are in effect