Add Functional Testing Debugging and FAQ sections.
Debugging section comes courtesy of the research @msssk did when there
was no guide such as this.

FAQ is split into several questions and hopefully will have more added
eventually.

Add Error Message FAQ

The page referred to for the JsonWireProtocol spec is marked as Obsolete
and links to the W3C WebDriver Spec instead. Not sure if this belongs in
a forthcoming PR.

Update documentation based on @dylans feedback.

According to feedback from @dylans, I have updated the wording in some
of these documentation additions. I have also updated the Modifier Keys
FAQ item to use Leadfoot instead of WD.js and removed the Deciphering
Errors FAQ item, as it appears to be unnecessary given Leadfoot's wrapping
of the WebDriver error reporting.