Fix wrong numFailedTests in Pretty
Currently we have the same treatment for suiteError which happens in `before` and `beforeEach`. But it causes the numFailedTests to be off, because with the error in `before`, that suite will run only 1 time and 1 fail will be recorded.
By adding `relatedTest` information to the suiteError for the `beforeEach` case, we can count the numFailedTests in the reporter correctly.