Coverage reporters can show uncovered files
Coverage reporters now support a 'sources' option, which is an array of
file glob patterns. If set, a Coverage reporter will add coverage
information for all files in 'sources' that don't already have coverage
data to the final report.

The files in sources should be the files that are actually run (post
transpiling). Source maps will be applied to hook the coverage data back
to original sources.

[ci skip]