Allow separate tunnel and webdriver proxy configs
If a top-level `proxy` config property is specified, it will be used for
the Leadfoot server. Otherwise, the Dig Dug tunnel's proxy value will be
used. Setting `proxy` to `null` or '' will force the Leadfoot server to
ignore the tunnel's proxy setting.

Similarly, if no tunnel proxy is specified by a config proxy is, the
config proxy will be used for the tunnel.

Resolves #509