Allow uncaught rejections/errors to be ignored
Provide two new config properties, warnOnUnhandledRejection and
warnOnUncaughtException. When one or both of these are set to true, any
unhandled rejection or exception will be treated as a warning rather
than a failing error. When set to a string or regexp, it will be matched
aagainst any error or rejection reason; matching errors/reasons will be
treated as warnings, others will be failures.

resolves #986