Fix handling of query args with empty values
URLSearchParams is used to initially parse the query string.
Unfortunately, it doesn't distinguish between `arg=` and `arg' both end
up as a key-value entry of `{ key: 'arg', value: '' }`. Intern
distinguishes between these -- `arg=` is an empty-valued arg, while
`arg` is a boolean true.

This commit adds some fixup code to the query handling that ensures that
query args with `=` are passed to the next level of parsing code with
the `=` still present.

fixes #991